fix: pr notes
